# Enable logger connector re-design