--- a/pytorch_lightning/core/lightning.py
+++ b/pytorch_lightning/core/lightning.py
@@ -381,7 +381,8 @@ class LightningModule(
         value = apply_to_collection(value, numbers.Number, self.__to_float_tensor)
 
         if self.trainer.logger_connector.should_reset_tensors(self._current_fx_name):
-            # when restarting an new epoch, reset the tensors
+            # if we started a new epoch (running it's first batch) the hook name has changed
+            # reset any tensors for the new hook name
             results.reset(metrics=False, fx=self._current_fx_name)
 
         results.log(