--- a/pytorch_lightning/trainer/evaluation_loop.py
+++ b/pytorch_lightning/trainer/evaluation_loop.py
@@ -88,8 +88,8 @@ class EvaluationLoop(object):
     def on_evaluation_start(self, *args: Any, **kwargs: Any) -> None:
         self.should_track_batch_outputs_for_epoch_end: bool = self._should_track_batch_outputs_for_epoch_end()
 
-        assert self.trainer.results is not None
-        self.trainer.results.to(device=self.trainer.lightning_module.device)
+        assert self.results is not None
+        self.results.to(device=self.trainer.lightning_module.device)
 
         if self.trainer.testing:
             self.trainer.call_hook('on_test_start', *args, **kwargs)