… CHANGELOG.md
… docs/source/extensions/logging.rst
… pytorch_lightning/trainer/properties.py
… pytorch_lightning/trainer/training_loop.py
… tests/callbacks/test_progress_bar.py
… tests/core/test_metric_result_integration.py
… tests/deprecated_api/test_remove_1-6.py
… tests/models/test_grad_norm.py
… tests/models/test_horovod.py
… tests/trainer/logging_/test_eval_loop_logging.py
… tests/trainer/logging_/test_logger_connector.py
… tests/trainer/logging_/test_train_loop_logging.py
… tests/trainer/loops/test_evaluation_loop.py
… tests/trainer/loops/test_evaluation_loop_flow.py
… tests/trainer/loops/test_training_loop_flow_scalar.py
… tests/trainer/optimization/test_multiple_optimizers.py