## What does this PR do?

Integrate the logger connector re-design with the loops.
Fix tests
Remove legacy tests

Part of #7631 

### pseudo-benchmark:

<details>

**MASTER**
----------
`ParityModuleMNIST`:
time: 0:01:14.02
profiler: 19.160 seconds `training_step`

`HeavyLoggingBoringModel`
memory: 18.3 MiB
time: 0:00:08.62
profiler: 3.298 seconds `training_step`

**Logging PoC**
---------------
`ParityModuleMNIST`:
time: 0:01:12.17
profiler: 20.911 seconds `training_step`

`HeavyLoggingBoringModel`
memory: 70.3 KiB
time: 0:00:06.86
profiler: 7.130 seconds `training_step`

</details>


<details>
<summary>Code</summary>

```python
import gc
import io
import pstats
import tracemalloc
from datetime import datetime

import torch
import torch.nn as nn
import torch.nn.functional as F
from torch.utils.data import DataLoader

from pytorch_lightning import Trainer, LightningModule
from pytorch_lightning.profiler import AdvancedProfiler
from tests import PATH_DATASETS
from tests.helpers import BoringModel, RandomDataset
from tests.helpers.datasets import MNIST


def collect_stats():
    gc.collect()
    snapshot = tracemalloc.take_snapshot()
    top_stats = snapshot.statistics("lineno")
    for stat in top_stats[:3]:
        print(stat)


class HeavyLoggingBoringModel(BoringModel):
    def __init__(self, memory=False):
        super().__init__()
        self.memory = memory

    def on_fit_start(self):
        if self.memory:
            tracemalloc.start(10)

    def on_fit_end(self):
        if self.memory:
            tracemalloc.stop()

    def training_step(self, batch, batch_idx):
        if self.memory and batch_idx % 50 == 49:
            collect_stats()

        loss = super().training_step(batch, batch_idx)["loss"]

        output_dict = {f"loss_{i}": loss for i in range(200)}
        self.log_dict(output_dict, on_step=True, on_epoch=True, prog_bar=True)

        return loss

    def train_dataloader(self):
        return DataLoader(RandomDataset(32, 500))


class ParityModuleMNIST(LightningModule):
    def __init__(self, memory=False):
        super().__init__()
        self.memory = memory
        self.c_d1 = nn.Linear(in_features=28 * 28, out_features=128)
        self.c_d1_bn = nn.BatchNorm1d(128)
        self.c_d1_drop = nn.Dropout(0.3)
        self.c_d2 = nn.Linear(in_features=128, out_features=10)

    def forward(self, x):
        x = x.view(x.size(0), -1)
        x = self.c_d1(x)
        x = torch.tanh(x)
        x = self.c_d1_bn(x)
        x = self.c_d1_drop(x)
        x = self.c_d2(x)
        return x

    def on_fit_start(self):
        if self.memory:
            tracemalloc.start(10)

    def on_fit_end(self):
        if self.memory:
            tracemalloc.stop()

    def training_step(self, batch, batch_idx):
        if self.memory and batch_idx % 50 == 49:
            collect_stats()

        x, y = batch
        y_hat = self(x)
        loss = F.cross_entropy(y_hat, y)
        self.log("loss", loss, on_step=True, on_epoch=True, prog_bar=True)
        return {"loss": loss}

    def configure_optimizers(self):
        return torch.optim.Adam(self.parameters(), lr=0.02)

    def train_dataloader(self):
        return DataLoader(MNIST(root=PATH_DATASETS, train=True, download=True), batch_size=2)


class MyAdvancedProfiler(AdvancedProfiler):
    def summary(self) -> str:
        recorded_stats = {}
        for action_name, pr in self.profiled_actions.items():
            if action_name != "training_step":
                continue
            s = io.StringIO()
            ps = pstats.Stats(pr, stream=s).strip_dirs().sort_stats("tottime")
            ps.print_stats(20)
            recorded_stats[action_name] = s.getvalue()
        return self._stats_to_str(recorded_stats)


def run(model_cls, test):
    print(f"{'=' * 30}\n {test} - {model_cls.__name__}\n{'=' * 30}")

    if test == "memory":
        trainer = Trainer(max_epochs=1, logger=False, progress_bar_refresh_rate=50, weights_summary=None)
        model = model_cls(memory=True)
        trainer.fit(model)

    elif test == "time":
        trainer = Trainer(max_epochs=1, logger=False, progress_bar_refresh_rate=50, weights_summary=None)
        model = model_cls()
        start = datetime.now()
        trainer.fit(model)
        end = datetime.now()
        print("Time: ", end - start)

    elif test == "profiler":
        trainer = Trainer(
            max_epochs=1,
            logger=False,
            profiler=MyAdvancedProfiler(),
            progress_bar_refresh_rate=50,
            weights_summary=None,
        )
        model = model_cls()
        trainer.fit(model)


if __name__ == "__main__":
    for model_cls in (ParityModuleMNIST, HeavyLoggingBoringModel):
        for test in ("memory", "time", "profiler"):
            if model_cls is ParityModuleMNIST and test == "memory":
                continue
            run(model_cls, test)
```

</details>

Recap:
- Negligible speed difference when using a real (aka medium-large+) model
- `training_step` takes longer as `self.log` does more now.
- Fixed memory
- Runtime is decreased (faster aggregation)

## Before submitting
- [x] Was this discussed/approved via a GitHub issue? (not for typos and docs)
- [x] Did you read the [contributor guideline](https://github.com/PyTorchLightning/pytorch-lightning/blob/master/.github/CONTRIBUTING.md), **Pull Request** section?
- [x] Did you make sure your PR does only one thing, instead of bundling different changes together?
- [x] Did you make sure to update the documentation with your changes? (if necessary)
- [x] Did you write any new necessary tests? (not for typos and docs)
- [x] Did you verify new and existing tests pass locally with your changes?
- [x] Did you update the [CHANGELOG](https://github.com/PyTorchLightning/pytorch-lightning/blob/master/CHANGELOG.md)? (not for typos, docs, test updates, or internal minor changes/refactorings)

## PR review

 - [x] Is this pull request ready for review? (if not, please submit in draft mode)
 - [x] Check that all items from **Before submitting** are resolved
 - [x] Make sure the title is self-explanatory and the description concisely explains the PR
 - [x] Add labels and milestones (and optionally projects) to the PR so it can be classified