@pep8speaks 2021-06-09
Hello @carmocca! Thanks for opening this PR. 









* In the file [`pytorch_lightning/trainer/training_loop.py`](https://github.com/PyTorchLightning/pytorch-lightning/blob/ce9b089bdbc1ca19ff9134e4f540d55b87a1c9ae/pytorch_lightning/trainer/training_loop.py):

> [Line 212:17](https://github.com/PyTorchLightning/pytorch-lightning/blob/ce9b089bdbc1ca19ff9134e4f540d55b87a1c9ae/pytorch_lightning/trainer/training_loop.py#L212): [W503](https://duckduckgo.com/?q=pep8%20W503) line break before binary operator























Do see the [Hitchhiker's guide to code style](https://goo.gl/hqbW4r)