@Queuecumber 2021-06-10
Could someone explain what the `batch_size` parameter does? I don't see it being used in the code anywhere and the docs don't explain it.